// types/foundry/common/data/data.mjs/itemData.d.ts
interface ItemDataConstructorData {
	/**
	 * The _id which uniquely identifies this Item document
	 * @defaultValue `null`
	 */
	_id?: string | null | undefined

	/**
	 * The name of this Item
	 */
	name: string

	/**
	 * An Item subtype which configures the system data model applied
	 */
	type: ItemDataSource["type"]

	/**
	 * An image file path which provides the artwork for this Item
	 * @defaultValue `ItemData.DEFAULT_ICON`
	 */
	img?: string | null | undefined

	/**
	 * The system data object which is defined by the system template.json model
	 * @defaultValue template from template.json for type or `{}`
	 */
	data?: DeepPartial<ItemDataSource["data"]> | null | undefined

	/**
	 * A collection of ActiveEffect embedded Documents
	 * @defaultValue `new EmbeddedCollection(ActiveEffectData, [], BaseActiveEffect.implementation)`
	 */
	effects?: ConstructorParameters<ConfiguredDocumentClass<typeof documents.BaseActiveEffect>>[0][] | null | undefined

	/**
	 * The _id of a Folder which contains this Item
	 * @defaultValue `null`
	 */
	folder?: InstanceType<ConfiguredDocumentClass<typeof documents.BaseFolder>> | string | null | undefined

	/**
	 * The numeric sort value which orders this Item relative to its siblings
	 * @defaultValue `0`
	 */
	sort?: number | null | undefined

	/**
	 * An object which configures user permissions to this Item
	 * @defaultValue `{ default: CONST.DOCUMENT_OWNERSHIP_LEVELS.NONE }`
	 */
	permission?: Partial<Record<string, foundry.CONST.DOCUMENT_OWNERSHIP_LEVELS>> | null | undefined

	/**
	 * An object of optional key/value flags
	 * @defaultValue `{}`
	 */
	flags?: ConfiguredFlags<"Item"> | null | undefined
}

type ItemDataBaseSource = PropertiesToSource<ItemDataBaseProperties>
type ItemDataProperties = ItemDataBaseProperties & ConfiguredData<"Item">
type ItemDataSource = ItemDataBaseSource & ConfiguredSource<"Item">